refactor: split the executor test module and factor the new dispatch sites
Two gates that `make gate` cannot reach — it performs no Rust compilation —
both tripped on this branch, and neither is satisfied by an allowance.

`make source-quality`: adding one import pushed
`executor/tests.rs` to 2501 lines, one over the ceiling. Split into a module
directory rather than granted an exception, along the section boundaries the
flat file already had: expressions, mutations, lists, strings, with the shared
helpers and imports at the module root. The `super::super::parser` paths the
flat file used are now an absolute import in `mod.rs`, so the submodules do
not encode their own depth.

The rest of the gate's complaints were the new `Clause::LoadCsv` arm growing
four functions that were already at their recorded ceilings. Each addition is
now factored out rather than inlined:

- `execute_single_clause` — the guard arm is one line against a named
  constant, and `FusedCountByType`'s 28-line operator body moved to its own
  method, leaving the dispatch reading as a table (329 -> 308 lines, 47 -> 45
  branches, nesting 3 -> 2).
- `parse_clause_sequence` — both soft-keyword clause heads (`LOAD CSV` and the
  pre-existing `FORMAT`) parse in their own methods; each owns a positional
  rule, so neither belongs inline (52 -> 49 branches, back under the limit).
- `parse_primary_expression_inner` — the parameter `.field` chain moved to
  `parse_parameter_property_chain` (back to its recorded 226/50).
- `validate_scope` — `UNWIND` and `LOAD CSV` are the only two clauses that
  originate rows and bind one variable, and the scope rule is identical, so
  they share one arm through `bind_row_source` instead of stating it twice.

`tests/api-baselines/source-quality.json` moves only downward: three metrics
tightened, and `execute_mutable_bounded`'s entry removed now that the LOAD CSV
refactor left it at 16 lines and within the defaults.

`cargo clippy --all-targets -- -D warnings`: two `#[cfg(test)]`
`KgliteBackend::new` call sites had not been recompiled by any package test
command, because they live in the bolt-server binary's test target. They now
pass the production Bolt default, `CsvImportPolicy::Denied`.
