+++ crates/kglite/src/graph/languages/cypher/parser/mod.rs
@@ -345,6 +345,37 @@ impl CypherParser {
     /// Returns the parsed clauses plus the trailing `OutputFormat` (only a
     /// top-level `FORMAT CSV` sets it to `Csv`; subquery bodies reject
     /// `FORMAT`).
+    /// Parse a leading `LOAD CSV`, or report the positional rule.
+    ///
+    /// Recognised even when misplaced, so a user who put it after a `MATCH`
+    /// gets the rule instead of `Unexpected token at start of clause:
+    /// Identifier("LOAD")`.
+    fn parse_leading_load_csv(&mut self, first: bool, in_subquery: bool) -> Result<Clause, String> {
+        if !first || in_subquery {
+            return Err(Self::misplaced_load_csv_error());
+        }
+        self.parse_load_csv_clause()
+    }
+
+    /// Parse the trailing `FORMAT <name>` marker. `FORMAT CSV` is the only
+    /// supported spelling, and it is rejected inside a `CALL { }` body.
+    fn parse_format_tail(&mut self, in_subquery: bool) -> Result<OutputFormat, String> {
+        if in_subquery {
+            return Err("FORMAT is not allowed inside a CALL { } subquery body".to_string());
+        }
+        self.advance(); // consume FORMAT
+        match self.peek() {
+            Some(CypherToken::Identifier(fmt)) if fmt.eq_ignore_ascii_case("CSV") => {
+                self.advance(); // consume CSV
+                Ok(OutputFormat::Csv)
+            }
+            other => Err(format!(
+                "Expected format name after FORMAT (supported: CSV), got {:?}",
+                other
+            )),
+        }
+    }
+
     pub(super) fn parse_clause_sequence(
         &mut self,
         end_at_rbrace: bool,
@@ -443,36 +474,16 @@ impl CypherParser {
                 Some(CypherToken::Foreach) => {
                     clauses.push(self.parse_foreach_clause()?);
                 }
-                // LOAD CSV — an external row source, legal only in leading
-                // position (see `load_csv`). Recognised even when misplaced so
-                // the user gets the positional rule instead of "Unexpected
-                // token at start of clause: Identifier(\"LOAD\")".
+                // The two soft-keyword clause heads. Both arrive as
+                // `Identifier` (neither word is reserved), and both own a
+                // positional rule, so each parses in its own method rather
+                // than inline — see `parse_leading_load_csv` and
+                // `parse_format_tail`.
                 Some(CypherToken::Identifier(_)) if self.identifier_opens_load_csv() => {
-                    if !clauses.is_empty() || end_at_rbrace {
-                        return Err(Self::misplaced_load_csv_error());
-                    }
-                    clauses.push(self.parse_load_csv_clause()?);
+                    clauses.push(self.parse_leading_load_csv(clauses.is_empty(), end_at_rbrace)?)
                 }
                 Some(CypherToken::Identifier(s)) if s.eq_ignore_ascii_case("FORMAT") => {
-                    if end_at_rbrace {
-                        return Err(
-                            "FORMAT is not allowed inside a CALL { } subquery body".to_string()
-                        );
-                    }
-                    // FORMAT CSV — must be last clause
-                    self.advance(); // consume FORMAT
-                    match self.peek() {
-                        Some(CypherToken::Identifier(fmt)) if fmt.eq_ignore_ascii_case("CSV") => {
-                            self.advance(); // consume CSV
-                            return Ok((clauses, OutputFormat::Csv));
-                        }
-                        other => {
-                            return Err(format!(
-                                "Expected format name after FORMAT (supported: CSV), got {:?}",
-                                other
-                            ));
-                        }
-                    }
+                    return Ok((clauses, self.parse_format_tail(end_at_rbrace)?))
                 }
                 Some(t) => {
                     return Err(format!("Unexpected token at start of clause: {:?}", t));
